utils: change the url / build behavior scheme
This is the first patch of the serie to differentiate builds and urls. As
of now we were using the --url option to specify a build-type.

Add a new option --build-type which specifies which type of build it is.
From this build-type if no url is specified, the url can be defaulted by
a configuration file.

This configuration file is specific per user, e.g. AGL, others...

Change-Id: I9ce801a7518b78ee859c6c3bbcad3a89e884e832
Signed-off-by: Loys Ollivier <lollivier@baylibre.com>