utils: define test-plans based on build-type
As the list of test is expending, we want to be able to specify test
plans grouping several test definitions. These test plans are build type
specific.

Create a config file that specifies the test plans for each build type.
If the user does not specify the tests to run then use the default ones
for the build-type.

Change-Id: Ied4e9f80d3e42787174189cd08499a2906e500ef
Signed-off-by: Loys Ollivier <lollivier@baylibre.com>