templates/machines: rpi3: switch to upstream device-type
Switch the rpi3 to use the upstream LAVA device-type.

Note that the the default kernel-command line arguments in the
upstream device-type are suited for the upstream kernel.  When
building with an yocto/AGL kernel, those need to be overridden, so
also update the machine template to use the correct kernel
command-line.

Change-Id: I823a1946bf761aa1019a9ec808fa99594337f4f4
Signed-off-by: Kevin Hilman <khilman@baylibre.com>